Image Projector app
New application that takes manual data entry of image name/URL pairs and displays a UI to select previously entered images.
The selected image is displayed in a separately launched window to be displayed in a separate monitor.
Entered data is stored in localStorage.
Entries are sorted by name but use the URLs as their ids, allowing you to edit the name of a URL and save it in place.